Refactor: voting reputation addons to work for all versions
